requirements.txt
# Core web framework and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6

# Database
aiosqlite>=0.19.0

# Audio processing
numpy>=1.24.0
soundfile>=0.12.1

# Speech-to-text - Both implementations for full functionality
faster-whisper>=0.9.0
openai-whisper>=20231117

# Text-to-speech
piper-tts>=1.2.0

# HTTP client
requests>=2.31.0

# System monitoring
psutil>=7.0.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Optional: GUI framework (if using tkinter GUI)
# tkinter is built-in to Python, no need to install

# Additional audio processing - Required for backend voice services
librosa>=0.10.1

# Machine learning - Required for VAD and advanced voice processing
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.35.0
# tensorflow>=2.13.0  # Optional - only if using TensorFlow models